requirements.txt
requests==2.20.0
idna==2.7
cryptography==2.3.1
pyOpenSSL==19.0.0
beautifulsoup4==4.7.1
progressbar2==3.39.3
